fix(vs): VSNetworkManager RttMs 改用 (int) cast，保持与 VampireSurvivors 目录一致

+++ unity/com.boom.boomnetwork/Samples~/VampireSurvivors/Scripts/Network/VSNetworkManager.cs
@@ -201,7 +201,7 @@ namespace BoomNetwork.Samples.VampireSurvivors
             else if (!wantsPause && _network.Client.IsGamePaused)
                 _network.Client.RequestGameResume();
 
-            _ui.UpdateHUD(_sim, _localSlot, _network.Client.RttMs);
+            _ui.UpdateHUD(_sim, _localSlot, (int)_network.Client.RttMs);
         }
 
         void OnDesync(FrameHashMismatch mismatch)
